Test: Remove getTopicindexNext from commons
